+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 Unreleased changes
 ------------------
 
+* ``km3pipe.plot``: Common plotting helpers
+
 Version 7
 ---------